refactor(fleet): move done_when parsing into its own module
loop_run.rs is 1201 lines, already past the 800-line rule, so the new
done-policy vocabulary lands beside done_marker in a module of its own
rather than growing it further. Adds DonePolicy and the queue-empty
sentinel; no behavior change yet.
